requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.4.2
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
Werkzeug==3.1.3
WTForms==3.2.1
